Add files via upload
What You're Aiming For
In this checkpoint, we are going to work on the 'Electric Vehicle Data' dataset that was provided by Kaggle as part of the Electric Vehicle Price Prediction competition.
Dataset description: This dataset contains information on the Battery Electric Vehicles (BEVs) and Plug-in Hybrid Electric Vehicles (PHEVs) that are currently registered with the Washington State Department of Licensing (DOL). This dataset was introduced as part of an official invitation-based competition on Kaggle. Our SVM model should answer the question "This is my car's model & make, along with a few other parameters, what price can this vehicle be brought or sold?”
➡️ Dataset link
https://i.imgur.com/IpuCW3s.jpg
➡️Columns explanation 
 
Instructions
1.	Import you data and perform basic data exploration phase
o	Display general information about the dataset
o	Create a pandas profiling reports to gain insights into the dataset
o	Handle Missing and corrupted values
o	Remove duplicates, if they exist
o	Handle outliers, if they exist
o	Encode categorical features
2.	Select your target variable and the features
3.	Split your dataset to training and test sets
4.	Build and train an SVM model on the training set
5.	Assess your model performance on the test set using relevant evaluation metrics
6.	Discuss with your cohort alternative ways to improve your model performance